Fix boolean properties, and punch up tests.
Review at http://gwt-code-reviews.appspot.com/755802

Review by: cromwellian@google.com

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@8649 8db76d5a-ed1c-0410-87a9-c151d255dfc7

+++ user/test/com/google/gwt/requestfactory/server/JsonRequestProcessorTest.java
@@ -115,7 +115,7 @@ public class JsonRequestProcessorTest extends TestCase {
       assertEquals(1, updateArray.getJSONObject(0).length());
       assertTrue(updateArray.getJSONObject(0).has("id"));
       assertFalse(updateArray.getJSONObject(0).has("violations"));
-      assertEquals((int) 45, (int) fooResult.getIntId());
+      assertEquals(45, (int) fooResult.getIntId());
       assertEquals("JSC", fooResult.getUserName());
       assertEquals(now, fooResult.getCreated());
       assertEquals(9L, (long) fooResult.getLongField());
@@ -148,7 +148,7 @@ public class JsonRequestProcessorTest extends TestCase {
       // check modified fields and no violations
       assertFalse(result.getJSONObject("sideEffects").has("UPDATE"));
       fooResult = SimpleFoo.getSingleton();
-      assertEquals((int) 45, (int) fooResult.getIntId());
+      assertEquals(45, (int) fooResult.getIntId());
       assertEquals("JSC", fooResult.getUserName());
     } catch (Exception e) {
       e.printStackTrace();
@@ -177,7 +177,7 @@ public class JsonRequestProcessorTest extends TestCase {
       // check modified fields and no violations
       assertTrue(result.getJSONObject("sideEffects").has("UPDATE"));
       fooResult = SimpleFoo.getSingleton();
-      assertEquals((int) 45, (int) fooResult.getIntId());
+      assertEquals(45, (int) fooResult.getIntId());
       assertEquals("JSC", fooResult.getUserName());
       assertEquals(newTime, fooResult.getCreated().getTime());
     } catch (Exception e) {
